+++ include/first_run/EN-SCOT.php
@@ -1,9 +1,11 @@
 <div class="row">
     <div class="col">
     <h2>Hello World</h2>
-    <p>
-        Hello, welcome to MyHomeBoardPHP <b>("MHBPHP")</b>, version <span style="color: green; font-weight: bold !important;"><?php print(MHBPHP_Version);?></span>
-    </p>
-
+        <p>
+            Hello, welcome to MyHomeBoardPHP <b>("MHBPHP")</b>, version <span style="color: green; font-weight: bold !important;"><?php print(MHBPHP_Version);?></span>.
+        </p>
+        <p>
+            Please update your configuration file. <pre>/include/configuration.php</pre>
+        </p>
     </div>
 </div>